[feat/#205] 월간 운동 건물 지도 데이터 조회 API 구현 (#210)
* feat: 컨트롤러 구현

* feat: DTO 구현

* feat: 날짜 및 위치 정보 생성

* feat: 운동 조회 로직 구현

* feat: 필요없는 모임 조인 제거

* feat: 날짜별 건물 그룹핑 로직 및 반환 로직 추가

* feat: Addr의 위도 및 경도 필드 Float에서 Double로 자료형 변경

* fix: longitude 대신 latitude를 보내고 있던 로직 에러 해결

* feat: 날짜를 입력하지 않으면 오늘을 기본값으로 변경

* docs: 스웨거 수정

* feat: 위도와 경도 입력하지 않을 시 기본값으로 멤버 메인 주소를 사용

* feat: 거리 계산 로직을 위해 멤버 주소 또한 Double로 변경 + 사이드 이펙트 처리

* fix: 생성된 주소값을 반환하도록 변경

* docs: api 경로 변경 및 스웨거 설명 변경

* refactor: 클래스명 및 변수명 명확화

* docs: 스웨거 설명 추가

+++ src/main/java/umc/cockple/demo/domain/member/dto/CreateMemberAddrDTO.java
@@ -18,9 +18,9 @@ public class CreateMemberAddrDTO {
 
             String buildingName,
             @NotNull(message = "위도는 필수입니다.")
-            Float latitude,
+            Double latitude,
             @NotNull(message = "경도는 필수입니다.")
-            Float longitude
+            Double longitude
 
     ) {
     }
